+++ fixtures/with-side-effects/src/entries/reset.ts
@@ -1,0 +1,7 @@
+import '../lib/reset';
+
+if (process.env.NO_SIDE_EFFECTS !== 'clearly') {
+  throw new Error('side-effect');
+}
+
+export * from '../lib/breakpoints';