+++ packages/core/src/plugins/rollup/externals.ts
@@ -1,13 +1,16 @@
 import path from 'path';
 
+import memoize from 'fast-memoize';
 import type { FunctionPluginHooks, Plugin } from 'rollup';
 import {
   externals as rollupExternals,
   type ExternalsOptions,
 } from 'rollup-plugin-node-externals';
-import type { PackageJson } from 'type-fest';
+import semverIntersects from 'semver/ranges/intersects';
 
+import type { EnhancedConfig } from '../../config';
 import { logger } from '../../logger';
+import type { PackageJson } from '../../types';
 import type { Format } from '../../types';
 import { promiseMap } from '../../utils/promise-map';
 import { resolveFrom } from '../../utils/resolve-from';
@@ -15,6 +18,32 @@ import { resolveFrom } from '../../utils/resolve-from';
 class PackagesById extends Map<string, PackageJson> {}
 
 const loadPackage = (packagePath: string): PackageJson => require(packagePath);
+
+const parseImportSpecifier = memoize((id: string) => {
+  let scope: string | undefined;
+  let scopedPackageId = id;
+  let packageId = id;
+  let isSubpath = false;
+
+  if (!id.startsWith('@') && id.includes('/')) {
+    isSubpath = true;
+    packageId = id.split('/')[0];
+    scopedPackageId = packageId;
+  }
+  if (id.startsWith('@') && Number(id.match(/\//g)?.length) > 1) {
+    isSubpath = true;
+    [scope, packageId] = id.split('/');
+    scopedPackageId = packageId;
+    packageId = `${scope}/${packageId}`;
+  }
+
+  return {
+    isSubpath,
+    packageId, // full package id, including the scope
+    scope,
+    scopedPackageId, // package id without the scope
+  };
+});
 
 async function loadPackageFrom(from: string, id: string): Promise<PackageJson> {
   try {
@@ -32,17 +61,17 @@ async function loadPackageFrom(from: string, id: string): Promise<PackageJson> {
   }
 }
 
-async function findDependencies(config: ExternalsOptions) {
-  const packagePath = config.packagePath! as string;
+async function findDependencies(options: ExternalsOptions) {
+  const packagePath = options.packagePath! as string;
   const packageRoot = path.dirname(packagePath);
 
   const packageJson = loadPackage(packagePath);
-  const externalDeps = [
-    ...Object.keys((config.deps && packageJson.dependencies) || {}),
-    ...Object.keys((config.devDeps && packageJson.devDependencies) || {}),
-    ...Object.keys((config.peerDeps && packageJson.peerDependencies) || {}),
-    ...Object.keys((config.optDeps && packageJson.optionalDependencies) || {}),
-  ];
+  const externalDeps = Object.keys({
+    ...(options.deps ? packageJson.dependencies : undefined),
+    ...(options.devDeps ? packageJson.devDependencies : undefined),
+    ...(options.peerDeps ? packageJson.peerDependencies : undefined),
+    ...(options.optDeps ? packageJson.optionalDependencies : undefined),
+  });
   const packagesById = new PackagesById();
 
   await promiseMap(externalDeps, async (dep) => {
@@ -53,23 +82,41 @@ async function findDependencies(config: ExternalsOptions) {
   return packagesById;
 }
 
-export function externals(packageRoot: string, format: Format = 'esm'): Plugin {
+export function externals(
+  config: EnhancedConfig,
+  format: Format = 'esm',
+): Plugin {
+  const packageRoot = config.root;
   const packagePath = path.join(packageRoot, 'package.json');
-  const config = {
+  const rootPackageJson = loadPackage(packagePath);
+  // preconstruct doesn't understand `satisfies`
+  const options: ExternalsOptions = {
     packagePath,
     deps: true,
     devDeps: false,
     peerDeps: true,
     optDeps: true,
   };
-  const plugin = rollupExternals(config);
+  const plugin = rollupExternals(options);
 
   let packagesById: PackagesById;
 
-  const patch = async (id: string) => {
+  const patchImportSpecifier = memoize(async (id: string) => {
     const resolvedId = await resolveFrom(packagePath, id);
-    return resolvedId.slice(resolvedId.lastIndexOf(id));
-  };
+    const { scope, scopedPackageId } = parseImportSpecifier(id);
+    return (
+      (scope ? `${scope}/` : '') +
+      resolvedId.slice(resolvedId.lastIndexOf(scopedPackageId))
+    );
+  });
+  const shouldReconcilePackage = memoize(
+    (packageId: string) =>
+      config.reconcileDependencies[packageId] &&
+      semverIntersects(
+        rootPackageJson.peerDependencies?.[packageId] ?? '*',
+        config.reconcileDependencies[packageId],
+      ),
+  );
 
   return {
     ...plugin,
@@ -79,7 +126,7 @@ export function externals(packageRoot: string, format: Format = 'esm'): Plugin {
     async buildStart(...args) {
       await Promise.all([
         (plugin as FunctionPluginHooks).buildStart.call(this, ...args),
-        findDependencies(config).then((result) => (packagesById = result)),
+        findDependencies(options).then((result) => (packagesById = result)),
       ]);
     },
 
@@ -97,23 +144,14 @@ export function externals(packageRoot: string, format: Format = 'esm'): Plugin {
           (typeof resolved === 'boolean' && !resolved) ||
           (typeof resolved === 'object' && resolved?.external)
         ) {
-          let packageId = id;
-          let isSubpath = false;
-
-          if (!id.startsWith('@') && id.includes('/')) {
-            packageId = id.split('/')[0];
-            isSubpath = true;
-          }
-          if (id.startsWith('@') && Number(id.match(/\//g)?.length) > 1) {
-            packageId = id.split('/').slice(0, 2).join('/');
-            isSubpath = true;
-          }
-
+          const { packageId, isSubpath } = parseImportSpecifier(id);
           const packageJson = packagesById.get(packageId);
           const patched = {
             id:
-              format === 'esm' && isSubpath && !packageJson?.exports
-                ? await patch(id)
+              format === 'esm' &&
+              isSubpath &&
+              (!packageJson?.exports || shouldReconcilePackage(packageId))
+                ? await patchImportSpecifier(id)
                 : id,
             external: true,
           };